patch seeder; patched creation of config; patched service rate fetch
